% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRoonDataIMS.R
\name{patRoonDataIMS}
\alias{patRoonDataIMS}
\alias{exampleDataPath}
\alias{exampleAnalysisInfo}
\title{Test IMS-HRMS data for the patRoon package}
\usage{
exampleDataPath(polarity = "positive", type = "ims")

exampleAnalysisInfo(polarity = "positive")
}
\arguments{
\item{polarity}{Either \code{"positive"} or \code{"negative"}.}
}
\description{
A small set of LC-IMS-HRMS data to test and demonstrate the functionality of patRoon.
}
\details{
This package contains data files from (1) a standard mixture containing various contaminants and (2) a blank. Both
were analysed in triplicate, using an Shimadzu Nexera UHPLC coupled to Bruker timsTOF Pro 2.0 using positive and
negative electrospray ionization. See also the \href{https://github.com/rickhelmus/patRoonData}{patRoonData} package
for example LC-HRMS data.

The complete data processing script can be found in data-raw/process.R

\code{exampleDataPath} is a helper function that returns the file path of the example data.

\code{exampleAnalysisInfo} is a helper function that generate analysis information that can be directly used
  by \pkg{patRoon}.
}
\author{
Jan C. Specker, Patricia Aguilar-Alarcón and Rick Helmus
}
