% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRoonDataIMS.R
\docType{data}
\name{suspects}
\alias{suspects}
\alias{suspectsPos}
\alias{suspectsNeg}
\title{Example suspect lists}
\format{
A \code{data.frame} with column \code{name}, \code{mz}, \code{rt} and \code{SMILES}, specifying the analyte
  name, \emph{m/z}, retention time (in seconds) and SMILES, respectively.

An object of class \code{data.frame} with 0 rows and 0 columns.

An object of class \code{data.frame} with 0 rows and 0 columns.
}
\usage{
suspectsPos

suspectsNeg
}
\description{
The \code{suspectsPos} and \code{suspectsNeg} datasets are example suspect lists for a set of environmental
contaminants (measured in positive and negative ionization, respectively). The datasets can be directly used with the
\code{\link[patRoon]{screenSuspects}} function.
}
\author{
Jan C. Specker, Patricia Aguilar-Alarcón and Rick Helmus
}
\keyword{datasets}
